Highlighting, settings, duplicates
