+++ src/PlaylistManager.cpp
@@ -4,6 +4,7 @@
 #include "PlaylistManager.hpp"
 #include "ResettableStaticPtr.hpp"
 #include "Settings.hpp"
+#include "Utils.hpp"
 
 #include <filesystem>
 #include <fstream>
@@ -17,54 +18,10 @@
 #include "System/Convert.hpp"
 #include "UnityEngine/Resources.hpp"
 #include "UnityEngine/ImageConversion.hpp"
-#include "UnityEngine/RenderTexture.hpp"
 #include "GlobalNamespace/CustomLevelLoader.hpp"
 #include "GlobalNamespace/CustomPreviewBeatmapLevel.hpp"
 
 using namespace RuntimeSongLoader;
-
-std::string GetBase64ImageType(std::string& base64) {
-    if(base64.length() < 3)
-        return "";
-    std::string sub = base64.substr(0, 3);
-    if(sub == "iVB")
-        return ".png";
-    if(sub == "/9j")
-        return ".jpg";
-    if(sub == "R0l")
-        return ".gif";
-    if(sub == "Qk1")
-        return ".bmp";
-    return "";
-}
-
-std::string SanitizeFileName(std::string fileName) {
-    std::string newName;
-    // yes I know not all of these are disallowed, and also that they are unlikely to end up in a name
-    static const std::unordered_set<unsigned char> replacedChars = {
-        '/', '\n', '\\', '\'', '\"', ' ', '?', '<', '>', ':', '*'
-    };
-    std::transform(fileName.begin(), fileName.end(), std::back_inserter(newName), [](unsigned char c){
-        if(replacedChars.contains(c))
-            return (unsigned char)('_');
-        return c;
-    });
-    if(newName == "")
-        return "_";
-    return newName;
-}
-
-bool UniqueFileName(std::string fileName, std::string compareDirectory) {
-    if(!std::filesystem::is_directory(compareDirectory))
-        return true;
-    for(auto& entry : std::filesystem::directory_iterator(compareDirectory)) {
-        if(entry.is_directory())
-            continue;
-        if(entry.path().filename().string() == fileName)
-            return false;
-    }
-    return true;
-}
 
 bool ShouldAddPack(std::string name) {
     if(currentFolder && folderSelectionState == 3) {
@@ -76,21 +33,6 @@ bool ShouldAddPack(std::string name) {
     return folderSelectionState == 0 || folderSelectionState == 2;
 }
 
-std::string GetLevelHash(GlobalNamespace::CustomPreviewBeatmapLevel* level) {
-    std::string id = STR(level->levelID);
-    // should be in all songloader levels
-    auto prefixIndex = id.find("custom_level_");
-    if(prefixIndex == std::string::npos)
-        return "";
-    // remove prefix
-    id = id.substr(prefixIndex + 14);
-    auto wipIndex = id.find(" WIP");
-    if(wipIndex != std::string::npos)
-        id = id.substr(0, wipIndex);
-    LOWER(id);
-    return id;
-}
-
 namespace PlaylistManager {
     
     std::unordered_map<std::string, Playlist*> name_playlists;
@@ -100,8 +42,6 @@ namespace PlaylistManager {
     std::hash<std::string> hasher;
     std::unordered_map<std::size_t, int> imageHashes;
     
-    // desired image size
-    const int imageSize = 256;
     // array of all loaded images
     std::vector<UnityEngine::Sprite*> loadedImages;
     // all unusable playlists
@@ -125,48 +65,6 @@ namespace PlaylistManager {
         "Custom Levels",
         "WIP Levels"
     };
-    
-    std::string ProcessImage(UnityEngine::Texture2D* texture, bool returnPngString) {
-        // check texture size and resize if necessary
-        int width = texture->get_width();
-        int height = texture->get_height();
-        if(width > imageSize && height > imageSize) {
-            // resize (https://gist.github.com/gszauer/7799899 modified for only downscaling)
-            auto texColors = texture->GetPixels();
-            ArrayW<UnityEngine::Color> newColors(imageSize * imageSize);
-            float ratio_x = ((float) width - 1) / imageSize;
-            float ratio_y = ((float) height - 1) / imageSize;
-
-            for(int y = 0; y < imageSize; y++) {
-                int offset_from_y = y * imageSize;
-
-                int old_texture_y = floor(y * ratio_y);
-                int old_texture_offset_from_y = old_texture_y * width;
-                
-                for(int x = 0; x < imageSize; x++) {
-                    int old_texture_x = floor(x * ratio_x);
-
-                    newColors[offset_from_y + x] = texColors[old_texture_offset_from_y + old_texture_x];
-                }
-            }
-            texture->Resize(imageSize, imageSize);
-            texture->SetPixels(newColors);
-            texture->Apply();
-        }
-        if(!returnPngString)
-            return "";
-        // convert to png if necessary
-        // can sometimes need two passes to reach a fixed result
-        // but I don't want to to it twice for every cover, so it will just normalize itself after another restart
-        // and it shouldn't be a problem through ingame cover management
-        auto bytes = UnityEngine::ImageConversion::EncodeToPNG(texture);
-        return STR(System::Convert::ToBase64String(bytes));
-    }
-
-    void WriteImageToFile(std::string_view pathToPng, UnityEngine::Texture2D* texture) {
-        auto bytes = UnityEngine::ImageConversion::EncodeToPNG(texture);
-        writefile(pathToPng, std::string(reinterpret_cast<char*>(bytes.begin()), bytes.Length()));
-    }
     
     int GetPackIndex(std::string title) {
         // find index of playlist title in config
@@ -405,8 +303,9 @@ namespace PlaylistManager {
                         // check for duplicate name
                         while(name_playlists.find(playlist->name) != name_playlists.end()) {
                             LOG_INFO("Duplicate playlist name!");
-                            RenamePlaylist(playlist, playlist->name + ".");
-                            path = playlist->path;
+                            playlist->name = playlist->name + ".";
+                            playlist->playlistJSON.PlaylistTitle = playlist->name;
+                            WriteToFile(path, playlist->playlistJSON);
                         }
                         name_playlists.insert({playlist->name, playlist});
                         path_playlists.insert({playlist->path, playlist});
